services/tts_requirements.txt
# services/tts_requirements.txt
pika==1.3.2
TTS==0.22.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
loguru==0.7.2
python-dotenv==1.0.1
pydantic-settings==2.3.4
spacy==3.7.5
sudachipy==0.6.8
sudachidict-core==20230110
